Adjusting how main is called in FPLBase
Because FPLBase has a dependency on SDL, the function main() of the
application has to include SDL_main.h, which will potentially modify
the function so that it will be called correctly, based on the
platform.

In order to hide this dependency, FPLBase will have a main function
declared, and users of the library will have to implement FPL_main(),
which will be called from FPLBase.

Change-Id: I5de29d8cff61cfcec5583b2aaa5f343e67d5f4bf