Create a pipeline to convert meshes from FBX to FlatBuffers.
When CMake option `fplbase_build_mesh_pipeline` is ON, the new
executible `mesh_pipeline` gets built into `bin`. This build
requires the FBX SDK, so by default it is OFF.

Also, fix a couple small things in the CMake builds.
(1) Libraries are now put into the `obj` directory, instead of
the main directory. (2) The warning about not having a
parent is now avoided.

Tested: Pipeline runs for simple meshes on Linux and Mac.
We do not build the pipeline on non-desktop platforms. It remains
to build and test the pipeline on Windows, and to write unit tests
for the pipeline, too.

Change-Id: I66b8297ac8926db4211f36463027f9f7b494762a