Separate texture, mesh, and material directories in pipeline.
We want to be able to export the texture and mesh files
to different directories, and reference the material (.fplmat)
from yet another directory.

Also, use the new generated functions like MeshExtension()
to get the extension from the schema files (.fbs) instead
of duplicating them in code.

Tested: Pipeline works in dependent projects, on Linux.
Change-Id: Ie0c72c4e5779de3d53a1f3902f0d428bad40bf6f