Keep local copy of vertex buffer definition inside Mesh.
This fixes a bug in MaterialManager::LoadMesh that prevented
meshes from dislaying. LoadMesh passed in an array that existed
on the stack, and Mesh kept a pointer to it. After LoadMesh
completed, that array got overwritten.

The fix is to keep a local copy of the array.

Also, fix copy-and-paste error in assert.

Tested: MaterialManager::LoadMesh works with our simple cube mesh.
Change-Id: I29cedeafa130be47e380eabe51e8de262edd7511