Rework mesh_pipeline to have unique uvs.
The new algorithm is simpler than the last and more
versatile. We loop through all poligon-vertices,
and add a new control point only if it doesn't match
an existing control point. We can still do this
in O(n) time by using a hash map to check if a
control point already exists.

Tested: Cube comes through fine on Linux and Mac.
Change-Id: I253cf399527577565445eea8d4057817d42168b0