Fix mesh_pipeline to use relative paths.
The paths are now relative to a base asset directory,
which is specified on the command line with the '-b' option.

All output files are now written to one directory.

We don't copy texture files anymore. The texture files
have to be processed by a separate pipeline anyway to be
put into the correct format, so we assume they're going to
end up in the right place.

We search more thoroughly for textures referenced by FBX files.
See comments in FindSourceTextureFileName for details.

Tested: Assets build correctly in test project on Linux.
Change-Id: I3cdef45bb383b65148cfc10178645391921158ec