Fixed an errant usage of GL_CALL.
The return value of GL_CALL was being used which is not supported.
GL_CALL is not always a simple wrapper; depending on which #defines are
set, it may be it's own block, so it can't be used in if or while
conditions. This change removes a usage of GL_CALL in an if condition,
and alters the definition of GL_CALL so that such an error is more
difficult to make in the future.

Tested the continuous build script on Linux.

Change-Id: I115a99eb92f6bd4375e7bc27034cde7508f2ded1