Allow emissive textures to be used as the base texture.
In Maya, emissive textures are those attached to the
"Incandescence" property of the material.

Also, improve logging of textures that are being ignored.

Log the shader index to which a texture is being assigned.

Tested: textures in sample app come through fine. FBX file
with emissive texture comes through fine.
Change-Id: Iaa0cc5288cbca71df8302c949bb951d5d4c69033