Output all textures attached to a mesh.
The immediate application of this is to use the specular map.

We now keep meshes separate if they have any textures that
are different. Previously we only compared the first texture
to see if they should be in the same surface.

Also cleaned-up the output messaging. By default, we output
only warnings and errors. There's a new -d option to output
important details. The -v option outputs all messages--this
part hasn't changed.

Tested: Works on all models in test app.
Change-Id: Ie144df86d50ca261336df4eb0ef7fb1c09c39fa9