Fix mesh_pipeline to output all surfaces, not just 1st.
If a mesh has different materials, then we split it into several
meshes. I assumed that each new mesh got a new node. However,
all of the meshes are actually attached to the same node.

The solution is to iterate through every mesh attached to a node.
We also have to be careful to only look at materials that
correspond to a given mesh.

Tested: Works on skybox fbx files, that have 5 different textures
on one mesh.
Change-Id: Ib59277f7b7817d654142d3223a49cc7b36c94bb8