Added a "ColorRGBA" primitive to the common flatbuffer types
Added "ColorRGBA" to the common flatbuffer definitions.  Also added
a loader for it, LoadColorRGBA, to match other flatbuffer loading
functions.  The ColorRGBA type and LoadColorRGBA are functionally
very similar to the Vec4 type and LoadVec4 function - the primary
differences are that colors have r, g, b, a properties (instead of
x, y, z ,w) and they default to 1.0 instead of 0.

Change-Id: Idfdd03b762c2c539fc6ae97d71ad3a78b342a6af