Export vertex colors in mesh_pipeline.
Vertex color is a vertex element in FBX. We collect the
vertex colors the same as UV coordinates, etc.

Vertex color can also come from the material as a
constant value. If there is no vertex color element,
and there are no textures, we use the constant value
from the material, if it has been set.

Tested: Works on sample scene with vertex colors and
another sample scene with colors only in the materials.
Running all meshes through the test app works, on Linux
and Mac.

Change-Id: I1760f0ef1f7b492df7307f379729b7ec9c9da93c