Fix normal and tangent calculations in mesh_pipeline.
If the mesh was offset, the normals would be incorrect.

The problem was that we were adding the translation onto
the normals and tangents. That is, we were transforming
them like points (w=1) instead of like vectors (w=0).

FBX's affine matrix class provides no way to multiply by
a vector, so we have to create a copy of the matrix with
the translation zeroed out.

Change-Id: I793008b6c068cdfc185064f32475d8c6af68de61
Tested: Test assets now come through with correct normals,
on Linux and Mac.