Add option to output mesh_pipeline hierarchy.
Enabled with -h option. By default, disabled.

Needed for animated meshes.

Vertices are output relative to their local pivots.
The transform from the parent pivot to the local pivot
is output in bone_transforms.

To contrast, by default the hierarchy is flattened
and all vertices are output relative to the root.

Change-Id: I5e5a43fea4c37b1ef408e55547910adcbea50556
Tested: Runs on test app.