Fixed OpenGL linker errors on Windows.
When building FPLBase on windows without SDL there was a series of
linker errors caused by how Windows treats OpenGL extensions.  On
Windows, all of the OpenGL extensions need to be declared as function
pointers rather than functions.  Removing the GL_GLEXT_PROTOTYPES define
removes the prototypes from being declared.

The other issue is that it seems that we were not declaring the
required function pointers anywhere as far as I can tell, so it still
doesn't link. In FPLBase when using SDL there is a block of code that
declares those function pointers. Removing the #ifdef around that code
allows those function pointers to be declared when using stdlib as well.

Tested on Linux and Windows.
Bug: 22489410

Change-Id: I1bbc3fd93e5e43449fd14ed7972954f641d998da