Fix bugs in hierarchy transform calculation.
There was a lot in the math that wasn't correct.

Also, added debug logging to verbose output. We output the
tranformation data for each node in the original scene
and our transformed scene. We also output the transformed
first vertex of every bone--which should be the same in
flat or hierarchical setups. Super-useful to debug the
hierarchy.

Tested: Test app assets run and look good on Mac and Linux.
Change-Id: I18a55658fd54c8123144cef4ac9a75c3d44ba2cb