mesh_pipeline now outputs intermediate nodes.
Previously we only output nodes with meshes, but all
nodes in the hierarchy can be animated.

Now we output all nodes that have meshes or children
with meshes.

Note that this change only has an affect when the '-h'
option (hierarchy) is specified. If there is no
hierarcy, then these intermediate nodes are irrelevent
because they don't have any vertices.

Change-Id: Ic452e253af405821f10420df16e6f61f4d9af42f
Tested: Works with sample game on Linux and Mac.