Upload bone transform to shader only if a vert indexes it.
Adds concept of `shader_transforms`, which is a subset of
the bone transforms. Not all bones have verts associated
with them. We still have to process them independently
because they may be animated. However, we don't need
to send those bone transforms to the vertex shader.

This reduces our vertex shader uniform usage by almost
half with the models in our test game.

Change-Id: I1236eb52a88f1e03c9531109abdf49de560ce204
Tested: Works on Linux and Android.