Fix crash on exit by properly destroying async loader.
The termination signal of a nullptr wasn't being sent to the worker
thread in the destructor, added the code to send the signal.

Sending the termination signal uncovered a bug where the worker
thread would never delete the nullptr from the work |queue_|,
resulting in a state where the TryFinalize() function would
never conclude that the worker thread was finished, because
the work |queue_| would never be empty. Repaired the logic
to erase the nullptr from the work |queue_| before exiting
the worker thread loop.

Change-Id: I0f3f6a233d9e22cbd9e79b061c41af5fc570883a