Fix Cardboard undistortion with hardware scaling.
If hardware scaling is used, the width x height will be less than the
displays natural resolution, so the PPI (pixels per inch) will also
be different. So, we use this trick to recalculate the ScreenParam's PPI
values (which are normally just read from the display).

Change-Id: Ifc1411590e638b02e80381254079b956fdf6ca3d
Tested: Works on Nexus 5