Allow texture file names to have different casing.
For example, if "BirdN.tga" is not found, we should also
look for "bird_n.tga" (i.e. camel case to snake case).

This is important because our external model get delivered
in casings that don't match our standard. It's error prone
and annoying to go through every model and change the
referenced texture names to match our case.

Much better to let the pipeline handle that.

Change-Id: I0c1464f1d62fcbc291ed71e6118cc6bb85f9b288
Tested: Works with camel_case names in our sample app, on Mac.