Reducing uniform usage in bone_transforms.
In order to reduce uniform usage, the shader has been changed to
use an array of vec4 instead of a mat4 for each 'bone_transform'.
In order to accommodate this change, the C++ shader code needed
to be changed to use vec4's as well.

Note: These files contain temporary code that is marked with TODOs
to be removed in the next commit. These exist because I need to
propagate the mat4-to-vec4 changes up to the 'renderer'. For now,
I simply accept mat4's from the 'renderer' and then manually convert
them to an array of vec4's. This is wasteful and should be removed
as soon as possible.

Bug: b/23494891
Change-Id: Ief72bbbcd7eb89d00103715e4830964d6899e9cd
Tested: This code was tested by compiling and running on OS X.