Add option to mesh_pipeline for distance unit.
The --unit option allows you to specify which unit of the
FBX file represents one unit in the game. Each FBX file
has a unit associated with it. This option converts the
mesh to the appropriate unit in the mesh_pipeline.

This makes the animator's job a lot easier, since the
animator no longer has to remember export settings.
Previously, the animator would have to remember that
one model was exported as cm and another as m. Now,
no matter what unit the model is exported in,
it will always end up with the same scale.

Change-Id: Idc5039aaf5dcf1133a62aecd8693b0ec34439f09
Tested: Works on sample app data.