Fixed a crash on shutdown.
This change includes a forked copy of SDL's SDLActivity in that includes
a minor change to fix a crash that occurs when exiting the app. When
exiting, under some circumstances a call to SDLSurface's surfaceUpdate
is made. This causes SDLSurface to attempt to start up another thread to
run native code. To prevent this a boolean tracking the exiting state
has been added.

Tested on Android.
Bug: 25386508
Change-Id: Iaa975dde6b69d79a5ab363ee85d921f3afa4e98c