Exposed the vsync ID via a getter function
Added a function to systrace which exposes the vsync frame id.  This
is a value that is guaranteed to change only during vsync events.
(Returns generally unique values, until it wraps.)  Useful for
determining if you've taken too long to render a frame and have
missed your vsync window.

Tested in zooshi builds android, and via make fplbase_stdlib.

Change-Id: I84a97d97201889e16debf163eacc362ffbeaa5a8