Fixing rotation to default to the same as Java
Device orientation was previously defaulting to 1 and querying Java
for the rotation at start up, but now Java updates the orientation when
it detects a change. The issue is if Java finds the rotation value to
be 0, it would not report it, and thus in the input would have the
rotation of 1, instead of 0.

Tested on Android, Nexus 5 and 10.

Change-Id: Ib53b9a24e69aca2727dfb8e47bb11be0cdfdcf4c