Adding checks if CardboardView is created properly
On some phones the CardboardView is failing to create properly, so
adding checks for that case, and having the check for supporting HMDs
depend on it being created properly. Also fixing a bug where that
support information was not propagating correctly.

Tested on Android, by temporarily having CardboardView not instantiate

Change-Id: Ifa61e86d59b3412155bc99b656a476d95db6a0d3