fplbase/cmake: add fplbase_use_external_sdl2
with external_sdl2_include_dir and external_sdl2_libraries for configuring
absolute path to include and compiled libraries.

This is useful for compiling fplbase against a pre-compiled version of
SDL installed with the host packaging system or built with a different
build system than cmake (ex: iOS/Xcode).

Tested on MacOSX with:
cmake -Dfplbase_use_external_sdl2=ON
-Dexternal_sdl2_include_dir=/usr/local/include/SDL2
-Dexternal_sdl2_lib_dir=/usr/local/lib
-Dexternal_sdl2_libraries="/usr/local/lib/libSDL2.dylib"  . && make

BUG:25284688
Change-Id: Iab9e52e81c63f934b032be0dc1a46ba348ccd989