Added Initialize() method for stdlib backend.
Initialize() currently only initializes the max number of vertex
uniforms but could be expanded in future to support other functionality
of the renderer that is not SDL specific.

Also, fixed API ref generation for the default (SDL) backend.

Tested:
Verified the project still continues to build on Linux (including the
fplbase_stdlib target).

Bug: 25458387
Change-Id: I0dcf1bb15c5e5f0a203413a9d334a184dc96a8aa