Cleaning up FPLBase docs.
The docs were rendering really oddly. The workaround seems to be
adding Modules (Groups) to the docs for each included file. For
some reason, doxygen does not generate the detail documentation
in the "File List" section alone.

I also added/fixed the documentation to a few areas that were
missing.

Bug: b/26111508

Tested: Tested on Ubuntu with generate_docs.py and viewed on
Chrome. I also build the library on Ubuntu to ensure I didn't
accidentally break anything.

Change-Id: Ifbd828d0a52ab070aca6efa67f17cd212958ee29