Fixing some warnings because "warnings as errors".
With the changes to our CMakeLists, we now use a common set of
compiler flags. As a result, there were two errors when compiling
for Linux.

`src/input_sdl.cpp` had a multi-line comment which was causing a
warning. Since it was the result of a `url`, the comment is now
on a single line that is greater than 80 characters (as allowed by
the exception in the Google Styleguide).

`src/utilities.cpp` was comparing the result of `fopen` from
`<cstdio>` with `0` instead of `NULL`. This was causing a warning
as well.

Tested: Tested on Linux.

Change-Id: I9039de44c8d0010dac2ba04fc4cd5819ee4ad63a