Fix min/max position calculation for skinned meshes.
We were using the bone-space values for the vertices in the
min/max calculation. We now use the object-space values.

Also, instead of keeping state to do the calculation, we
do the calculation only when we need it. It's not all that
much more inefficient, and it is one less moving part.

Tested: Works on hippo model.
Change-Id: Ic8cc19b64491dc5f8adc26568a0ab9f406d530ac