Fix mesh_pipeline build.
The mesh_pipeline sets CMAKE linker and cxx flags slightly
differently. This was messing up the samples and tests builds
that came *after* the mesh_pipeline build.

The fix is to move the mesh_pipeline build code into its own
CMakeLists.txt file so that it can't wreck the common variables
for the other builds.

Also, we now call append_flags() on CMAKE_CXX_FLAGS instead of
stomping the variable entirely. This ensures that common flags,
such as for ignoring unused functions (mesh_pipeline didn't have
this so it was spitting out warnings), don't get stomped.

Change-Id: I1de97f1dda9f14d7ea5253964358c18c178c53d5
Tested: Works on Linux and Mac.