Output vertices in object space instead of bone space.
Previously, vertices were output in bone space so that they could
be more easily animated. However, that created a host of
animation problems, and it also made it more difficult to display
the mesh (since identity transforms for bones would not give the
default pose).

Depricate the --hierarchy flag. It does nothing now.

Tested: Works on Zooshi on Linux and Mac.
Change-Id: I381b075ef9418b78286d74c3cd1f6ab4a4a3b003