Added cubemap support.
Should work for all texture types, including compressed textures.
Works by simply creating a 1x6 ratio texture, and specifying in
the material file (or Texture/LoadTexture) that this is supposed
to be a cubemap.

Simple test by loading it into the mesh sample, could do something
more spectacular with it at some point.

Bug: 29646856
Change-Id: I63d4edb59fdc80b4771142d4452ca8c2a38066e2