Hack SDL's CMakeLists.txt file so it works with cmake 3.5 and up.
SDL's CMakeLists.txt file has a check to ensure it's not being
built in-tree. However, their theck is too strict. We *never* build
SDL in-tree, but sometimes we build fplbase in-tree. The SDL check
fails if the top-most project is being built in tree.

This change makes SDL fail only if SDL itself is being built in-tree.

Obviously, hacking SDL's CMakeLists.txt file at build time is not a
long term solution. We have a bug open with them here,
https://bugzilla.libsdl.org/show_bug.cgi?id=3269

Bug: 27255873
Change-Id: I460366c8e0c90f0b354c57c22389548f7a4d0bfd
Tested: Works on both cmake 2.8.12.1 and 3.5.2 for OSX.