Fix fplbase break on Linux with older versions of cmake.
The file(LOAD) command is broken on older versions of cmake.
It will remove semicolons on file load. See this bug for details,
https://cmake.org/pipermail/cmake/2006-January/008043.html.

The (even more hacky) fix is to keep a local version of SDL's
CMakeLists.txt and copy it over before the add_subdirectory()
call. This more fragile, but avoids the file(LOAD) call.

Change-Id: Ia4e213206643fc59def02a5092ae8231054f073e
Tested: Works on Linux cmake 2.8.12.1.