Add a version number to the fplmesh data format.
This will ensure that we'll fail early when trying to load stale
mesh data. Breaking format changes are detectable even though
FlatBuffers is resilient with legacy data.

Note that existing data is unaffected by this change. Existing
data with have `Unspecified` version, which is explicitly ignored
on load.

We also need to have `Unspecified` since not all meshes are created
by the mesh_pipeline. Some simple meshes (for the ground or skybox)
are specified in json files. We don't want to have to update the
version on each of these simple json files every time we make a
change to unrelated mesh features.

Tested: Zooshi data fails to load when version number incremented
but data is not regenerated.

Change-Id: I56a58ebbaf9a6c2126aa0343d925bfee764f5cd7