Add support for a second set of UVs to mesh_pipeline.
Useful for light-maps, for example.

Also, fix the logic for determining which vertex attributes
to output. By default, it is supposed to output only the
vertices that exist in the FBX file, but we were always
outputing tangents and normals no matter what (we'd generate
them if they didn't exist). The new behavior only generates
tangents and normals if they've been explicitly requested
with the --attrib switch.

Bug: 30501110
Tested: Test file with two UV sets loads in games.
Change-Id: I880b3b6f849a0075d6991ddaedf4b1f58851f535