mesh_pipeline now gracefully handles more than 2^16 verts and 2^8 bones.
Previously we'd just assert. In release builds, we would output a file
that had bad vert indices or bad bone data.

Now for verts, we output an error, truncate to the max number of verts
we can index, and remove any polys that use verts beyond that index.

For bones we weigh to the root bone (0) when indices exeed 255.

Neither of these options solve the problem, but they are the best that
we can do given current runtime limitations.

Tested: Ran on scene with many polys, many vertex attributes, and many bones.
Change-Id: I76b6d256144b75d41557b4dd12b9c97edd88c58a