Fix obj loading in mesh_pipeline.
Just needed to remove early exit when not an FBX file.

Also removed stale code that output an error when >2^16 verts.
We properly handle that case now. The error was just left-over.

Also improve the speed of TextureFromNode() by gathering the unique
material indices before processing them. There was a material index
per-face in a large obj, and even though every face had the same index,
we were doing the same processing over-and-over.
PiperOrigin-RevId: 142263222
Change-Id: I95da0bdd11373438b8b6eae2c040d4f3fc7d2697