Get fplbase samples building for Android.
Needed to replicate some build functionality from fplbase/jni/Android.mk and fplbase/jni/sdl/Android.mk.

Also needed to fix some basic compile errors in SDL2's BUILD file (e.g. the define for GL_GLEXT_PROTOTYPES), and to add the SDL2_jni_onload_renamed target, which lets callers hook into the JNI_OnLoad() function.

Removed the Cardboard functionality from FPLActivity. This SDK has been replaced and is no longer available internally. We have a task to upgrade FPLActivity to the latest VR SDK. When we do, we should probably leave FPLActivity as is, and derive from it an FPLVRActivity.

FPLActivity now overrides SDLActivity.getLibraries() so that we don't have to (1) create a dummy libSDL2.so, or (2) rename our native library main.so. We assume that our library is called lib[app_name].so, where app_name is a string defined in the Android resource file.

It remains to update the Android.mk files.
PiperOrigin-RevId: 144107804
Change-Id: I3c1f2f875dbb48e9bb34cac2712749beafc78bca