Soft skinning support for FPL mesh_pipeline and fplviewer.
The most significant change is in mesh_pipeline.cpp, where it now tracks multiple bone weights and indices.  Additionally, it traverses the entire scene to gather skeleton bones, rather than just nodes containing meshes.  This is necessary because soft skinning uses non-hierarchical binding, so meshes influenced by a bone won't necessarily be a descendant of that bone.

This effectively changes the skeleton, breaking compatibility with anim_pipeline.  So this change also updates anim_pipeline.cpp to use a compatible scene traversal.

For testing, I also changed the skinned vertex shader in viewer.cpp to use all 4 weights.  Other projects using this shader will have to be updated in another CL, but the mesh_pipeline changes are compatible with rigid skinning so it shouldn't break existing content.
PiperOrigin-RevId: 146259013
Change-Id: I4dc33f1fc0f0e3bb79600b6e08addc7944ec9978