Automated g4 rollback of changelist 162942622.
*** Reason for rollback ***

Doesn't compile on OSX -- _EAC and _ETC2 formats not defined.

*** Original change description ***

Limit the number of compressed mips by the block size.  On Android O, trying to load a mip that is smaller than the block size generates GL_INVALID_OPERATION.  This also fixes a Linux issue which was previously addressed by an additional call to glGenerateMipmaps.

Tested on linux and android O in a sample app using textureLod.

***

PiperOrigin-RevId: 162952442
Change-Id: I795e97564a3455f5a7bb6a38a322396355d9b565